add scribejava and discord implementation

+++ src/r3qu13m/mei/launcher/core/ModExtractor.java
@@ -6,11 +6,15 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
-import java.security.MessageDigest;
+import java.util.Map;
+import java.util.Optional;
+import java.util.UUID;
 
 import org.apache.commons.codec.digest.DigestUtils;
 
-import r3qu13m.mei.lib.MeiLogger;
+import r3qu13m.mei.lib.MPVec;
+import r3qu13m.mei.lib.MeiServerLib;
+import r3qu13m.mei.lib.OperationType;
 import r3qu13m.mei.lib.structure.DistributeFile;
 import r3qu13m.mei.lib.structure.Mod;
 import r3qu13m.mei.lib.structure.ModPack;
@@ -40,7 +44,9 @@ public class ModExtractor {
 	}
 
 	private static void deleteFile(final File target) {
-		target.renameTo(new File(target.getAbsolutePath() + ".bak"));
+		if (target.exists()) {
+			target.renameTo(new File(target.getAbsolutePath() + ".bak"));
+		}
 	}
 
 	private static String computeHash(final File target) {
@@ -84,10 +90,27 @@ public class ModExtractor {
 				distribute.getName(), sha1, distribute.getHash()));
 	}
 
-	public static void extract(final File baseDir, final ModPack pack) {
+	public static void extract(final File baseDir, final ModPack pack, final Optional<ModPack> currentPack) {
+		Map<UUID, OperationType> diff;
+		if (currentPack.isPresent()) {
+			diff = new MPVec(pack, currentPack.get()).getDifference();
+		} else {
+			diff = new MPVec(pack).getDifference();
+		}
+
+		for (UUID key : diff.keySet()) {
+			if (diff.get(key) == OperationType.DELETE) {
+				ModExtractor.deleteFile(new File(ModExtractor.getTempDir(baseDir),
+						MeiServerLib.instance().getDistributeFile(key).getName()));
+			}
+		}
+
 		for (Mod mod : pack.getMods()) {
 			for (DistributeFile distribute : mod.getFiles()) {
-				File downloaded = ModExtractor.downloadFile(baseDir, distribute);
+				final UUID id = distribute.getID();
+				if (diff.containsKey(id) && diff.get(id) == OperationType.ADD) {
+					ModExtractor.downloadFile(baseDir, distribute);
+				}
 			}
 		}
 	}
